Added extra string handling for pydantic models to avoid having relationships with inexistent nodes
