+++ examples/using_hnsw_as_vectorDB.py
@@ -2,12 +2,13 @@ import os
 from openai import AsyncOpenAI
 from dotenv import load_dotenv
 import logging
-
+import numpy as np
+from sentence_transformers import SentenceTransformer
 from nano_graphrag import GraphRAG, QueryParam
 from nano_graphrag._llm import gpt_4o_mini_complete
 from nano_graphrag._storage import HNSWVectorStorage
 from nano_graphrag.base import BaseKVStorage
-from nano_graphrag._utils import compute_args_hash
+from nano_graphrag._utils import compute_args_hash, wrap_embedding_func_with_attrs
 
 logging.basicConfig(level=logging.WARNING)
 logging.getLogger("nano-graphrag").setLevel(logging.DEBUG)
@@ -15,6 +16,19 @@ logging.getLogger("nano-graphrag").setLevel(logging.DEBUG)
 WORKING_DIR = "./nano_graphrag_cache_using_hnsw_as_vectorDB"
 
 load_dotenv()
+
+
+EMBED_MODEL = SentenceTransformer(
+    "sentence-transformers/all-MiniLM-L6-v2", cache_folder=WORKING_DIR, device="cpu"
+)
+
+
+@wrap_embedding_func_with_attrs(
+    embedding_dim=EMBED_MODEL.get_sentence_embedding_dimension(),
+    max_token_size=EMBED_MODEL.max_seq_length,
+)
+async def local_embedding(texts: list[str]) -> np.ndarray:
+    return EMBED_MODEL.encode(texts, normalize_embeddings=True)
 
 
 async def deepseepk_model_if_cache(
@@ -77,6 +91,7 @@ def insert():
         cheap_model_max_async=10,
         best_model_func=deepseepk_model_if_cache,
         cheap_model_func=deepseepk_model_if_cache,
+        embedding_func=local_embedding
     )
     start = time()
     rag.insert(FAKE_TEXT)
@@ -95,7 +110,7 @@ def query():
         cheap_model_max_async=4,
         best_model_func=gpt_4o_mini_complete,
         cheap_model_func=gpt_4o_mini_complete,
-        
+        embedding_func=local_embedding
     )
     print(
         rag.query(
